requirements.txt
Pillow==7.1.1
pyTelegramBotAPI==3.6.7
requests==2.23.0
six==1.14.0
comtypes==1.1.7